move interpreter-hooks work to vm-rework
