+++ external-crates/move/move-vm/runtime/src/lib.rs
@@ -26,7 +26,5 @@ mod tracing;
 #[cfg(any(debug_assertions, feature = "debugging"))]
 mod debug;
 
-mod paranoid_type_checker;
-mod plugin;
 #[cfg(test)]
 mod unit_tests;
